refactor(c++): iterative structure changed.

+++ c++/practices/iterative-control-structure/pi-value.cpp
@@ -4,16 +4,18 @@ using namespace std;
 
 int main(){
     
-    float inputNumber, piValue = 4;
+    double inputNumber, piValue = 4;
 
     cout << endl << "[========= PI VALUE =========]" << endl << endl;
 
     cout << "Enter the number of iterations: ";
     cin >> inputNumber;
 
-    for (float i = 3; i <= (inputNumber * 4) ; i += 4) {
+    for (double i = 3; i <= (inputNumber * 4) ; i += 4) {
         piValue = piValue - (4/i) + (4/(i+2));
     }
+    
     cout << "The value of pi is: " << piValue << endl;
+    
     return 0;
 }
